+++ Tests/GRDBTests/RowFromDictionaryLiteralTests.swift
@@ -105,29 +105,32 @@ class RowFromDictionaryLiteralTests : RowTestCase {
         assertRowConvertedValueEqual(row, column: Column("c"), value: CustomValue.c)
     }
     
-    func testDataNoCopy() {
+    func testWithUnsafeData() throws {
         do {
             let data = "foo".data(using: .utf8)!
             let row: Row = ["a": data]
             
-            XCTAssertEqual(row.dataNoCopy(atIndex: 0), data)
-            XCTAssertEqual(row.dataNoCopy(named: "a"), data)
-            XCTAssertEqual(row.dataNoCopy(Column("a")), data)
+            try row.withUnsafeData(atIndex: 0) { XCTAssertEqual($0, data) }
+            try row.withUnsafeData(named: "a") { XCTAssertEqual($0, data) }
+            try row.withUnsafeData(at: Column("a")) { XCTAssertEqual($0, data) }
+            
+            try row.withUnsafeData(named: "missing") { XCTAssertNil($0) }
+            try row.withUnsafeData(at: Column("missing")) { XCTAssertNil($0) }
         }
         do {
             let emptyData = Data()
             let row: Row = ["a": emptyData]
             
-            XCTAssertEqual(row.dataNoCopy(atIndex: 0), emptyData)
-            XCTAssertEqual(row.dataNoCopy(named: "a"), emptyData)
-            XCTAssertEqual(row.dataNoCopy(Column("a")), emptyData)
+            try row.withUnsafeData(atIndex: 0) { XCTAssertEqual($0, emptyData) }
+            try row.withUnsafeData(named: "a") { XCTAssertEqual($0, emptyData) }
+            try row.withUnsafeData(at: Column("a")) { XCTAssertEqual($0, emptyData) }
         }
         do {
             let row: Row = ["a": nil]
             
-            XCTAssertNil(row.dataNoCopy(atIndex: 0))
-            XCTAssertNil(row.dataNoCopy(named: "a"))
-            XCTAssertNil(row.dataNoCopy(Column("a")))
+            try row.withUnsafeData(atIndex: 0) { XCTAssertNil($0) }
+            try row.withUnsafeData(named: "a") { XCTAssertNil($0) }
+            try row.withUnsafeData(at: Column("a")) { XCTAssertNil($0) }
         }
     }